メンバー非表示機能追加

+++ database/migrations/2025_03_13_144525_add_attendance_id_to_records_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('records', function (Blueprint $table) {
-            // attendance_id カラムを追加
-            $table->unsignedBigInteger('attendance_id');
+        if (!Schema::hasColumn('records', 'attendance_id')) {
+            Schema::table('records', function (Blueprint $table) {
+                // attendance_id カラムを追加
+                $table->unsignedBigInteger('attendance_id');
 
-            // 外部キーの設定
-            $table->foreign('attendance_id')
-                  ->references('id')
-                  ->on('attendances')
-                  ->onDelete('cascade'); // attendanceが削除されたら、それに紐づくrecordsも削除
-        });
+                // 外部キーの設定
+                $table->foreign('attendance_id')
+                    ->references('id')
+                    ->on('attendances')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -29,10 +31,12 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('records', function (Blueprint $table) {
-            // 外部キー制約を削除
-            $table->dropForeign(['attendance_id']);
-            // attendance_id カラムを削除
-            $table->dropColumn('attendance_id');
+            if (Schema::hasColumn('records', 'attendance_id')) {
+                // 外部キー制約を削除
+                $table->dropForeign(['attendance_id']);
+                // カラム削除
+                $table->dropColumn('attendance_id');
+            }
         });
     }
 };
